Changes request validator interface into an abstract base class to save repetition
