+++ src/PinguApps.Appwrite.Shared/Requests/CreateAccountRequest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.Json.Serialization;
+using PinguApps.Appwrite.Shared.Requests.Validators;
 using PinguApps.Appwrite.Shared.Utils;
 
 namespace PinguApps.Appwrite.Shared.Requests;
@@ -6,7 +7,7 @@ namespace PinguApps.Appwrite.Shared.Requests;
 /// <summary>
 /// The request for creating an account
 /// </summary>
-public class CreateAccountRequest
+public class CreateAccountRequest : BaseRequest<CreateAccountRequest, CreateAccountRequestValidator>
 {
     /// <summary>
     /// User ID. Choose a custom ID or generate a random ID with <see cref="IdUtils.GenerateUniqueId(int)"/>. Valid chars are a-z, A-Z, 0-9, period, hyphen, and underscore. Can't start with a special char. Max length is 36 chars
@@ -32,3 +33,5 @@ public class CreateAccountRequest
     [JsonPropertyName("name")]
     public string? Name { get; set; }
 }
+
+
