+++ src/PinguApps.Appwrite.Shared/Requests/UpdatePhoneRequest.cs
@@ -1,13 +1,12 @@
 ﻿using System.Text.Json.Serialization;
-using FluentValidation;
-using FluentValidation.Results;
+using PinguApps.Appwrite.Shared.Requests.Validators;
 
 namespace PinguApps.Appwrite.Shared.Requests;
 
 /// <summary>
 /// The request for updating a users phone
 /// </summary>
-public class UpdatePhoneRequest : IRequestValidator
+public class UpdatePhoneRequest : BaseRequest<UpdatePhoneRequest, UpdatePhoneRequestValidator>
 {
     /// <summary>
     /// Phone number. Format this number with a leading '+' and a country code, e.g., +16175551212
@@ -20,27 +19,5 @@ public class UpdatePhoneRequest : IRequestValidator
     /// </summary>
     [JsonPropertyName("password")]
     public string Password { get; set; } = string.Empty;
-
-    /// <inheritdoc/>
-    public bool IsValid() => Validate().IsValid;
-
-    /// <inheritdoc/>
-    public ValidationResult Validate(bool throwOnFailures = false)
-    {
-        var validator = new UpdatePhoneRequestValidator();
-
-        if (throwOnFailures)
-            return validator.Validate(this, x => x.ThrowOnFailures());
-
-        return validator.Validate(this);
-    }
 }
 
-internal class UpdatePhoneRequestValidator : AbstractValidator<UpdatePhoneRequest>
-{
-    public UpdatePhoneRequestValidator()
-    {
-        RuleFor(x => x.Phone).NotEmpty().Matches("^\\+\\d*$");
-        RuleFor(x => x.Password).NotEmpty().MinimumLength(8);
-    }
-}
