+++ src/PinguApps.Appwrite.Shared/Requests/Validators/UpdatePhoneRequestValidator.cs
@@ -1,0 +1,11 @@
+﻿using FluentValidation;
+
+namespace PinguApps.Appwrite.Shared.Requests.Validators;
+public class UpdatePhoneRequestValidator : AbstractValidator<UpdatePhoneRequest>
+{
+    public UpdatePhoneRequestValidator()
+    {
+        RuleFor(x => x.Phone).NotEmpty().Matches("^\\+\\d*$");
+        RuleFor(x => x.Password).NotEmpty().MinimumLength(8);
+    }
+}
